+++ src/rstr/compkit.cpp
@@ -67,7 +67,7 @@ extern c_comp wcomp;
 extern MN * main_number_ptr;
 extern BOX * boxchain, *dl_last_in_chain;
 extern uchar lpool[];
-extern uint16_t lpool_lth;
+extern uint lpool_lth;
 extern uchar work_raster[];
 extern uchar records_change;
