+++ src/rstr/tools.cpp
@@ -83,7 +83,7 @@ static cell first_cell, last_cell;
 static cell * empty_cell, *free_cell_chain, *cell_boundary;
 extern c_comp wcomp;
 extern CIF::version * start_rec, *rec_ptr;
-extern int16_t lpool_lth;
+extern int lpool_lth;
 extern uchar lpool[];
 
 static puchar save_kit(puchar c, int16_t l);
